fixed error page issue, need to debug direction switching for video analysis

+++ main.py
@@ -160,6 +160,7 @@ def determine_right(results,sensitivity=0.75):
 if __name__ == '__main__':
     path = 'Videos/satvik_erg.mp4'
     cap = cv2.VideoCapture(0)
+    print("reading camera")
     fps = cap.get(cv2.CAP_PROP_FPS)
     mpPose = mp.solutions.pose
     mpDraw = mp.solutions.drawing_utils
@@ -184,6 +185,9 @@ if __name__ == '__main__':
     init_frame = 0
     filler = np.ones((1024,1024,3), np.uint8) * 255
     error_message = False
+    has_not_destroyed = True
+    finish_shoulder_pos = 0
+    catch_shoulder_pos = 0
     while True:
         curr_alpha = 0.275 # alphas[a_idx]
         ended_stroke = False
@@ -204,27 +208,32 @@ if __name__ == '__main__':
         if results.pose_landmarks:
             min_shoulder_pos = min(results.pose_landmarks.landmark[shoulder].x,min_shoulder_pos)
             max_shoulder_pos = max(results.pose_landmarks.landmark[shoulder].x,max_shoulder_pos)
+            print("running poses")
             # Add check here (stall with continues)
             if not valid_detection and determine_right(results) == -1:
-                erorr_message = True
+                error_message = True
                 cv2.putText(filler, "Please move back so your body is visible", (200,512), cv2.FONT_HERSHEY_SIMPLEX, 1, (0,0,0), 2)
                 cv2.imshow("placeholder", filler)
                 cv2.waitKey(1)
+                print("stuck")
                 continue
             elif determine_right(results) == 0 and not valid_detection:
-                print("ready to rumble!!")
                 FLIP_CODE = 1
                 valid_detection = True
+                print("stuck again")
                 continue
             else:
-                print("ready to rumble!!")
                 valid_detection = True
-            if error_message:
-                print("attempted destruction")
+            if error_message and has_not_destroyed:
+                print("tried to break")
                 cv2.destroyWindow("placeholder")
+                print("broken?")
+                has_not_destroyed = False
+
            # mpDraw.draw_landmarks(img, results.pose_landmarks,mpPose.POSE_CONNECTIONS)
             h,w,c = img.shape
             for id, lm in enumerate(results.pose_landmarks.landmark):
+                print("continuing")
                 cx,cy = int(lm.x*w),int(lm.y*h)
                 if id in prev_positions_ema:
                     px,py = prev_positions_ema[id]
@@ -242,10 +251,12 @@ if __name__ == '__main__':
                                 # adjust this based on glitches
                                 if pos_locations_dict[shoulder][-1][0] > pos_locations_dict[shoulder][-7][0]:
                                     finish_frame = frame_idx
+                                    finish_shoulder_pos = pos_locations_dict[shoulder][-1][0]
                                     finish = True
                             elif id == 11 and len(pos_locations_dict[shoulder]) >= 7:
                                 if pos_locations_dict[shoulder][-1][0] < pos_locations_dict[shoulder][-7][0]:
                                     finish_frame = frame_idx
+                                    finish_shoulder_pos = pos_locations_dict[shoulder][-1][0]
                                     finish = True
                     elif finish and not catch:
                         pos_locations_dict[id].append((cx,cy))
@@ -256,10 +267,12 @@ if __name__ == '__main__':
                                 # adjust this based on glitches
                                 if pos_locations_dict[shoulder][-1][0] < pos_locations_dict[shoulder][-7][0]:
                                     catch_frame = frame_idx
+                                    catch_shoulder_pos = pos_locations_dict[shoulder][-1][0]
                                     catch = True
                             elif id == 11:
                                 if pos_locations_dict[shoulder][-1][0] > pos_locations_dict[shoulder][-7][0]:
                                     catch_frame = frame_idx
+                                    catch_shoulder_pos = pos_locations_dict[shoulder][-1][0]
                                     catch = True
                 prev_positions_ema[id] = (cx,cy)
             for i,f in right_connections:
@@ -274,7 +287,6 @@ if __name__ == '__main__':
             init_frame = frame_idx
             ended_stroke = True
             catch = False
-
             finish = False
             waited = 0
             # print(f"Distance traveled by shoulder: {round(max_shoulder_pos-min_shoulder_pos,3)*100} % of width")
@@ -284,17 +296,18 @@ if __name__ == '__main__':
             # if stroke_count > 1:
                 # compare_ref(user_list,ref_stroke)
             j_data = None
-            j_path = 'user_strokes.json' #TODO: make this more robust (so it doesn't crash if poses are missing)
-            if os.path.getsize(j_path) > 0:
-                with open(j_path, 'r') as inputs:
-                    j_data = json.load(inputs)
-            else:
-                with open(j_path, 'w') as output:
-                    json.dump([], output)
-                    j_data = []
-            j_data.append(pos_locations_dict)
-            with open('user_strokes.json', 'w') as output:
-                json.dump(j_data, output, indent=2)
+            j_path = 'user_strokes.json'
+            if abs(finish_shoulder_pos - catch_shoulder_pos) > 700:
+                if os.path.getsize(j_path) > 0:
+                    with open(j_path, 'r') as inputs:
+                        j_data = json.load(inputs)
+                else:
+                    with open(j_path, 'w') as output:
+                        json.dump([], output)
+                        j_data = []
+                j_data.append(pos_locations_dict)
+                with open('user_strokes.json', 'w') as output:
+                    json.dump(j_data, output, indent=2)
             # TODO: Add comparison function here
